separated event_bus module
