+++ src/edge_device/device_controller.py
@@ -1,34 +1,11 @@
+# device_controller.py
 import asyncio
 import logging
 from src.mqtt.aws_mqtt import IotCoreBrokerConnection
 from src.serial.serial_communication import SerialCommunication
+from src.edge_device.event_bus import EventBus
 
 logger = logging.getLogger(__name__)
-
-class EventBus:
-    def __init__(self):
-        self.listeners = {}
-        logger.debug("EventBus initialized with an empty listeners dictionary.")
-
-    def subscribe(self, event_type, listener):
-        if event_type not in self.listeners:
-            self.listeners[event_type] = []
-            logger.debug(f"New event type '{event_type}' added to listeners.")
-        self.listeners[event_type].append(listener)
-        logger.debug(f"Listener added for event type '{event_type}'.")
-
-    async def publish(self, event_type, data):
-        if event_type in self.listeners:
-            logger.debug(f"Publishing data to listeners of event type '{event_type}'.")
-            for listener in self.listeners[event_type]:
-                if asyncio.iscoroutinefunction(listener):
-                    await listener(data)
-                    logger.debug(f"Async listener called for event type '{event_type}'.")
-                else:
-                    listener(data)
-                    logger.debug(f"Sync listener called for event type '{event_type}'.")
-        else:
-            logger.debug(f"No listeners found for event type '{event_type}'. No action taken.")
 
 class DeviceController:
     def __init__(self, config):
@@ -111,6 +88,8 @@ class DeviceController:
             await self.serial_comm.disconnect()
 
     async def on_gnss_data_received(self, data):
+        # TODO edit this code so that is publishes to the message type specific topic
+        # $aws/rules/gnss_data_basic_ingest/{data['message_identity']}/self.
         aws_basic_ingest_rule_for_gnss_data = f"$aws/rules/gngga_basic_ingest/{self.config.ts_domain_name}/{self.config.device_id}"
         logger.debug(f"sending gnss data to {aws_basic_ingest_rule_for_gnss_data}")
         logger.debug(f"Received serial data: {data} ready to send to mqtt broker")
